Implement action search
This commit ditches g_desktop_app_info_search() with a modified copy of
the search code in order to support searching for actions in desktop
entries. That additionally required implementing own sorting, which
currently is based on the specificity of the match type (e.g. name ranks
above display name and keywords and so on) and whether we’ve
prefix-matched.

Resolves #5.